backend/requirements.txt
# Core
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0
gevent==22.10.2

# Security
Flask-Talisman==1.0.0
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7  # For PostgreSQL

# Monitoring
prometheus-flask-exporter==0.22.4
blinker==1.6.2

# Utils
requests==2.31.0
python-json-logger==2.0.7
python-multipart==0.0.6